Fix infinite scroll.
Structure of the scrollable::State has changed. Tuple type
matching the structure of the scrollable::State is past to
second param of mem::transmute. The third parameter of this tuple
(the last param of the scrollable::State struct) is what we want
to modify. So now tmp.2 is modified instead of tmp.1.